* conv_creds.c (krb5int_encode_v4tkt): Unsigned vs signed warning fix
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@16360 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/krb/conv_creds.c
@@ -191,7 +191,8 @@ int krb5int_encode_v4tkt(v4tkt, buf, encoded_len)
      buflen = *encoded_len;
 
      if (v4tkt->length < MAX_KTXT_LEN)
-	  memset(v4tkt->dat + v4tkt->length, 0, MAX_KTXT_LEN - v4tkt->length);
+	  memset(v4tkt->dat + v4tkt->length, 0, 
+		 (unsigned int) (MAX_KTXT_LEN - v4tkt->length));
      temp = v4tkt->length;
      if ((ret = encode_int32(&buf, &buflen, &temp)))
 	  return ret;
